add completion handler support via option.onComplete
This commit adds an options.onComplete option, which is called after all files have been downloaded and extracted. To achieve this I had to slightly modify when the downloadUrl callback is called, i.e., it is now passed along as a callback to extract, rather than being called immediately after extract is invoked (which, because extract makes subsequent asynchronous calls, was pretty much immediately).